with stylesheet, need to reset form still

+++ meme.js
@@ -1,44 +1,38 @@
 "use strict";
 
+let memes = [];
 
-let memes = [
-   
-];
 function showMemes() {
-    console.log("showMemes");
-
-    let memeArea = document.getElementById("memes");
+    let memeArea = document.getElementById("memes"); 
     let html = "";
 
     for(let i = 0; i < memes.length; i++){
         let meme = memes[i];
-        let img = "<img src=\" " +meme.imgUrl +" \">";
-        let topText = "<h3>" + meme.topText + "</h3>";
-        let bottomText = "<h3>" + meme.bottomText + "</h3>";
+        let img = "<img src=\" " + meme.imgUrl +" \">";
+        let topText = "<h3 class=topText>" + meme.topText + "</h3>";
+        let bottomText = "<h3 class=bottomText>" + meme.bottomText + "</h3>";
+        let deleteButton = "<button id= \"delete-" + i + "\" class = \"del\">X</i></button>";
         html +=
-            "<div id= \"meme-" + i +"\"" + "class= \"meme\">" +"\n"
+            "<div id=\" +i + \" class= \"meme\">" +"\n"
              + topText + "\n"
              + img + "\n"
              + bottomText + "\n"
-            + "</div>;"
+             + deleteButton 
+            + "</div>";
     }
 
     memeArea.innerHTML = html;
 }
 
 function addMeme(imgUrl, topText, bottomText) {
-    console.log("addMeme", imgUrl, topText, bottomText);
-
     memes.push({
         imgUrl: imgUrl,
         topText: topText,
         bottomText: bottomText,
     });
-    
 }
 
 function handleSubmit(evt) {
-    console.log('handleSubmit', evt);
     evt.preventDefault();
 
     let form = document.getElementById('meme-form');
@@ -51,7 +45,6 @@ function handleSubmit(evt) {
 
 function enterKey(evt) {
     if (evt.key === "Enter"){
-    console.log('enterKey', evt);
     evt.preventDefault();
 
     let form = document.getElementById('meme-form');
@@ -62,24 +55,20 @@ function enterKey(evt) {
     showMemes();}
 
 }
-
-
+function deleteMeme(id){
+    memes.splice (id, 1);
+}
 
 function handleDelete(evt){
-    console.log("handleDelete", evt);
-    let memeId = Number(evt.target.id.slice(5));
-    deleteMeme(memeId);
-    showMemes();
+    if (evt.target.id.startsWith("delete-")){
+        let memeId = Number(evt.target.id.replace("delete-", ""));
+        deleteMeme(memeId);
+    }
+showMemes();
 }
 
-function deleteMeme(id){
-    console.log("delete Meme", id)
-    memes.splice(id, 1)
-}
 
 showMemes();
-
-
 
 document.getElementById("meme-form").addEventListener("submit", handleSubmit);
 document.getElementById("meme-form").addEventListener("keypress", enterKey);
